Add fetchCurrentUser method to AuthService
This commit adds a new method called  to the  class in the  file. The method retrieves the current user information by making an HTTP GET request to the  endpoint, using the JWT token stored in the local storage for authorization. The retrieved user data is then used to authenticate the user in the application.